updated type and namespace colors
this is minded on cpp, so it would have to be checked against other
filetypes, but the namespace color is very offputting, and the only
color defined inline, which makes me think it was a mistake.

I'm creating another PR here, because this part is more opinionated.

The reason I also changed the type color, is also due to the way C++
is structured. A lot of things can happen inside of template<here> so it
being the same color as keywords and operators, makes it one big
homogenous blob